refactor: 뉴스스탠드 롤링 CLASS 적용
