--- src/components/news-stand-rolling/main.js
@@ -1,8 +1,0 @@
-import { LatestNewsSubject } from './subject.js';
-import { LatestNewsObserver } from './observer.js';
-
-const latestNewsSubject = new LatestNewsSubject();
-const latestNewsObserver = new LatestNewsObserver(latestNewsSubject);
-latestNewsSubject.addObserver(latestNewsObserver);
-
-export default latestNewsSubject;
