--- src/components/news-stand-rolling/observer.js
@@ -1,13 +1,0 @@
-import { Observer } from '../../utils/observer.js';
-
-class LatestNewsObserver extends Observer {
-  constructor(subject) {
-    super(subject);
-  }
-
-  update() {
-    this.subject.initLatestNews();
-  }
-}
-
-export { LatestNewsObserver };
