+++ src/index.js
@@ -1,13 +1,10 @@
 import newsStandHeader from './components/news-stand-header.js';
-
-import latestNewsSubject from './components/news-stand-rolling/main.js'; //
+import latestNews from './components/news-stand-latest.js';
 import pressHeaderStore from './components/press-header.js';
 
 (function init() {
   newsStandHeader.initHeader();
-
-  //
-  latestNewsSubject.initNews();
+  latestNews.initNews();
 
   // initPressHeader(); // 수정중
   pressHeaderStore.initPress();
